[sil] Move SILParser into its own local header.
This will let me add functionality into the SILParser into other files. The main
file is already very large.
